SoizeGhanemFactory: Hide SoizeGhanemFactorEval
